requirements.txt
asgiref==3.5.2
asttokens==2.0.8
backcall==0.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.5
cryptography==40.0.2
decorator==5.1.1
defusedxml==0.7.1
Django==3.2.13
django-allauth==0.54.0
django-extensions==3.2.0
django-seed==0.3.1
djangorestframework==3.13.1
executing==1.0.0
Faker==18.4.0
idna==3.4
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.6
oauthlib==3.2.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.30
psycopg2==2.9.6
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.2.1
requests==2.29.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.5.0
toposort==1.10
traitlets==5.3.0
urllib3==1.26.15
wcwidth==0.2.5
